+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Fixed a bug where it wasn’t possible to access the Edit Plan page.
 - Fixed a bug where it wasn’t possible to select entries from a different site when editing a discount.
 - Fixed a bug where it was possible to create another store when there were still sales.
+- Fixed a bug where soft deleted purchasables were 
 
 ## 5.0.0-beta.3 - 2024-04-10